+++ openapi-diff/src/modeler/AutoRest.Swagger/ComparisonContext.cs
@@ -3,6 +3,7 @@ using OpenApiDiff.Core;
 using OpenApiDiff.Core.Logging;
 using System;
 using System.Collections.Generic;
+using AutoRest.Swagger.Model;
 
 namespace AutoRest.Swagger
 {
@@ -13,7 +14,7 @@ namespace AutoRest.Swagger
     public class ComparisonContext<T>
     {
         private readonly JsonDocument<T> _CurrentRootDoc;
-        private readonly JsonDocument<T> _PreviousRootDoc; 
+        private readonly JsonDocument<T> _PreviousRootDoc;
 
         /// <summary>
         /// Initializes a top level context for comparisons