+++ src/lib/util/logging.ts
@@ -14,12 +14,12 @@ let logDir: unknown
  */
 function getTimeStamp() {
   // We pad each value so that sorted directory listings show the files in chronological order
-  function pad(number: any) {
+  function pad(number: number) {
     if (number < 10) {
       return "0" + number
     }
 
-    return number
+    return number.toString()
   }
 
   const now = new Date()
@@ -77,7 +77,7 @@ Object.defineProperties(log, {
       }
       const validLevels = Object.keys(customLogLevels)
       if (
-        !validLevels.some(function(item) {
+        !validLevels.some(function (item) {
           return item === level
         })
       ) {